requirements.txt
pandas==2.2.0
numpy==1.26.3
python-Levenshtein==0.23.0